Refactor: ショップのレーティングを表示するコードをShopRatingコンポーネントとして分割
